pydantic - clean up OpjectBase
